do not use both setParent(None) and deleteLater()

+++ pychoacoustics/UML_method.py
@@ -121,7 +121,7 @@ def setupUML(model="Logistic", swptRule="Up-Down", nDown=2, centTend="Mean", sti
     elif UML["par"]["x0"] == UML["par"]["x"]["limits"][1]:
         UML["current_step"] = UML["nsteps"]
     else:
-        UML["current_step"] = np.ceil(UML["nsteps"]/2)
+        UML["current_step"] = int(np.ceil(UML["nsteps"]/2))
 
 
     return UML
